dra structured: clarify construction of alloc.requestData
The comment about "all mode" wasn't attached to the code handling it
anymore. Moved.

Constructing entries with subrequests was interleaved with
non-subrequests. Reordered, renamed variables for consistency.

Removed the validateDeviceRequest hasSubRequests parameter because it is always
false.

Modify the requestKey while adding additional subrequest entries because then
the difference is easier to see.