DRA API: revise partitionable devices validation
For nested maps, the overall number of leaf entries is a better estimate of the
overall size. DRA driver authors then have more flexibility how to balance
number of sets vs. numbers of counters per set, without validation having to
make the individual limit for each unnecessarily large.

For the new entry in the Device struct, we should avoid making that struct
larger in the worst case. This can be achieved by counting attributes,
capacities and the new counters together: if counters are not used, the limit
is the same as before. If they are used, drivers know about the restriction if
we apply it together with introducing the new field.

While at it, make one "empty map" check a bit more resilient.